+++ app/posts/[id]/client.tsx
@@ -1,18 +1,15 @@
+'use client'
 import { NextSeo } from 'next-seo'
 import ReactMarkdown from 'react-markdown'
-import {Prism as SyntaxHighlighter} from 'react-syntax-highlighter'
-import { a11yDark } from 'react-syntax-highlighter/dist/cjs/styles/prism';
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter'
+import { a11yDark } from 'react-syntax-highlighter/dist/cjs/styles/prism'
 import remarkGfm from 'remark-gfm'
-import Layout from '../../components/layout'
+import Layout from '../../../components/layout'
 import rehypeRaw from 'rehype-raw'
-import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
-import { useAppDispatch } from 'reducers/store'
-import { changeLanguage } from 'reducers/locale_slice'
+import { useAppDispatch } from '../../../reducers/store'
+import { changeLanguage } from '../../../reducers/locale_slice'
 import { useEffect } from 'react'
-import { i18n } from 'next-i18next.config'
-import DefaultSeo from '../../next-seo.config'
-import { GetServerSideProps } from 'next'
-import GithubAPI from '@/lib/githubAPI'
+import DefaultSeo from '../../../next-seo.config'
 import 'gitalk/dist/gitalk.css'
 import Gitalk from 'gitalk'
 
@@ -26,42 +23,37 @@ interface Issue {
 }
 
 interface PostProps {
-  locale: typeof i18n.locales[number]
+  locale: string
   post: Issue
 }
 
-export default function Post (props: PostProps): React.ReactElement {
-  const {
-    post,
-    locale
-  } = props
+export default function PostClient(props: PostProps) {
+  const { post, locale } = props
   const dispatch = useAppDispatch()
   useEffect(() => {
-    dispatch(changeLanguage(locale))
-
+    dispatch(changeLanguage(locale as any))
     const gitalk = new Gitalk({
       clientID: process.env.NEXT_PUBLIC_GITALK_CLIENT_ID as string,
       clientSecret: process.env.NEXT_PUBLIC_GITALK_CLIENT_SECRET as string,
       repo: 'comments',
       owner: 'jhang-jhe-wei',
-      admin: ['jhang-jhe-wei']
+      admin: ['jhang-jhe-wei'],
     })
-
     gitalk.render('gitalk-container')
-  }, [])
+  }, [locale])
 
   return (
     <>
       <NextSeo
         title={post.title}
         description={post.description}
-        canonical={`https://wells.tw/posts/${post.title}}`}
+        canonical={`https://wells.tw/posts/${post.title}`}
         openGraph={{
           ...DefaultSeo.openGraph,
           locale,
           url: 'https://wells.tw/blog',
           title: post.title,
-          description: post.description
+          description: post.description,
         }}
       />
       <Layout>
@@ -75,14 +67,10 @@ export default function Post (props: PostProps): React.ReactElement {
                 rehypePlugins={[rehypeRaw]}
                 components={{
                   img: ({ node, ...props }) => (
-                    <img
-                      {...props}
-                      width={800}
-                      height={800}
-                    />
+                    <img {...props} width={800} height={800} />
                   ),
-                  code: (props) => {
-                    const {children, className, node, ...rest} = props
+                  code: props => {
+                    const { children, className, ...rest } = props as any
                     const match = /language-(\w+)/.exec(className || '')
                     return match ? (
                       <SyntaxHighlighter
@@ -92,50 +80,23 @@ export default function Post (props: PostProps): React.ReactElement {
                         language={match[1]}
                         wrapLines={true}
                         style={a11yDark}
-                        customStyle={{
-                          backgroundColor: 'transparent',
-                        }}
+                        customStyle={{ backgroundColor: 'transparent' }}
                       />
                     ) : (
-                        <code {...rest} className={className}>
-                          {children}
-                        </code>
-                      )
-                  }}
-                }
-                children={post.body}
-              />
-            <div id="gitalk-container"></div>
+                      <code {...rest} className={className}>
+                        {children}
+                      </code>
+                    )
+                  },
+                }}
+              >
+                {post.body}
+              </ReactMarkdown>
+              <div id="gitalk-container"></div>
             </article>
           </div>
         </main>
       </Layout>
     </>
   )
-}
-
-export const getServerSideProps: GetServerSideProps<PostProps> = async (context) => {
-  const {
-    locale,
-    params: { id }
-  } = context
-
-  const result = await GithubAPI.request('GET /repos/{owner}/{repo}/issues/{issue_number}', {
-    owner: 'jhang-jhe-wei',
-    repo: 'jhang-jhe-wei.github.com',
-    issue_number: id,
-    headers: {
-      'X-GitHub-Api-Version': '2022-11-28'
-    }
-  })
-
-  return {
-    props: {
-      locale,
-      post: result.data,
-      ...(await serverSideTranslations(locale, [
-        'common'
-      ]))
-    }
-  }
 }
