+++ next.config.js
@@ -2,6 +2,9 @@ const { i18n } = require('./next-i18next.config')
 
 module.exports = {
   i18n,
+  experimental: {
+    appDir: true
+  },
   eslint: {
     // Warning: This allows production builds to successfully complete even if
     // your project has ESLint errors.
